backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pillow
sqlalchemy
aiosqlite
langchain
openai
python-dotenv
pdf2image
aiofiles
asyncpg
greenlet